"proper table relations seeting"

+++ Services/CharacterService/CharacterService.cs
@@ -36,10 +36,14 @@ namespace dotnet_rpg.Services.CharacterService
     {
         var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
         Character character = _mapper.Map<Character>(newCharacter);
+        character.User = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
 
         _context.Characters.Add(character);
         await _context.SaveChangesAsync();
-        serviceResponse.Data = await _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();//adding srvcerespns data to character
+        serviceResponse.Data = await _context.Characters
+                
+                .Where(c => c.User.Id == GetUserId())
+                .Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();//adding srvcerespns data to character
         return serviceResponse; //returns ServiceResponse object  as response
     }
 
@@ -48,19 +52,28 @@ namespace dotnet_rpg.Services.CharacterService
         var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
         try
         {
-            Character character = await _context.Characters.FirstAsync(c => c.Id == id);
-            _context.Characters.Remove(character);
-            await _context.SaveChangesAsync();
-            
-            serviceResponse.Data = _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+            Character character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
+                if (character != null)
+                {
+                    _context.Characters.Remove(character);
+                    await _context.SaveChangesAsync();
 
+                    serviceResponse.Data = _context.Characters
+                        .Where(c => c.User.Id == GetUserId())
+                        .Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+                }
+                else
+                {
+                    serviceResponse.Success = false;
+                    serviceResponse.Message = "Character not found.";
+                }
         }
         catch (Exception ex) 
         {
             serviceResponse.Success = false;
             serviceResponse.Message = ex.Message;
         }
-        return serviceResponse; 
+        return serviceResponse;
     }
 
 
@@ -75,7 +88,7 @@ namespace dotnet_rpg.Services.CharacterService
     public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
     {
         var serviceResponse = new ServiceResponse<GetCharacterDto>();
-        var dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
+        var dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
         serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter); //serviceResponse.Data is actual result
         return serviceResponse;
 
@@ -84,28 +97,35 @@ namespace dotnet_rpg.Services.CharacterService
     public async Task<ServiceResponse<GetCharacterDto>> UpdateCharacter(UpdateCharacterDto updatedCharacter)
     {
         var serviceResponse = new ServiceResponse<GetCharacterDto>();
-        
-        try
-        {
-        Character character= await _context.Characters.FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
 
-        character.Name = updatedCharacter.Name;
-        character.Hitpoints = updatedCharacter.Hitpoints;
-        character.Strenght = updatedCharacter.Strenght;
-        character.Defence = updatedCharacter.Defence;
-        character.Inteligence = updatedCharacter.Inteligence;
-        character.Class = updatedCharacter.Class;
-        
-        await _context.SaveChangesAsync();
+            try
+            {
+                Character character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
+                if (character.User.Id == GetUserId())
+                {
+                    character.Name = updatedCharacter.Name;
+                    character.Hitpoints = updatedCharacter.Hitpoints;
+                    character.Strenght = updatedCharacter.Strenght;
+                    character.Defence = updatedCharacter.Defence;
+                    character.Inteligence = updatedCharacter.Inteligence;
+                    character.Class = updatedCharacter.Class;
 
-        serviceResponse.Data = _mapper.Map<GetCharacterDto>(character);
+                    await _context.SaveChangesAsync();
 
-        }
-        catch (Exception ex) 
-        {
-            serviceResponse.Success = false;
-            serviceResponse.Message = ex.Message;
-        }
+                    serviceResponse.Data = _mapper.Map<GetCharacterDto>(character);
+
+                }
+                else
+                {
+                    serviceResponse.Success = false;
+                    serviceResponse.Message = "Character not found.";
+                }
+            }
+            catch (Exception ex)
+            {
+                serviceResponse.Success = false;
+                serviceResponse.Message = ex.Message;
+            }
         return serviceResponse;
     }
